Fix sample usage of tag helper, and nonce helper
The current example in the sample site demonstrates the usage of the 2
helpers, however, in the configured CSP unsafe inline and eval are
allowed, so the examples work regardless of the proper working of the
helpers.

Further, the .WithHashTagHelper(), and .WithNonce() weren't invoked so
hashes, and the nonce weren't added to the CSP header.

This makes the example work properly.